organise menu and pages
