+++ stockpile-web/src/components/common/Navigation.js
@@ -7,10 +7,10 @@ const Navigation = ({brand}) => {
             <NavBarItem id="navHome" label="Home" link="/" active="true"></NavBarItem>
             <NavBarItem id="navArticle" label="Article" link="/articles"></NavBarItem>
             <NavBarItem id="navUpload" label="Upload" link="/data-import"></NavBarItem>
-            <NavBarItem id="navAbout" label="About" link="#">
-                <NavBarItem id="navConfigureUploadSchema" label="Configure upload schema" link="#" />
-                <NavBarItem id="navAboutUs" label="About" link="/about" />
+            <NavBarItem id="navAdmin" label="Admin" link="#">
+                <NavBarItem id="navConfigureUploadSchema" label="Configure upload schema" link="/admin/configure-upload-schemas" />
             </NavBarItem>
+            <NavBarItem id="navAboutUs" label="About" link="/about" />
         </NavBar>
     )
 }
