+++ stockpile-web/src/components/pages/admin/UploadSchemas.js
@@ -1,0 +1,10 @@
+
+const UploadSchemas = () => {
+    return (
+        <div>
+            hello
+        </div>
+    )
+}
+
+export default UploadSchemas
