+++ stockpile-web/src/components/pages/article/ArticleListing.js
@@ -1,5 +1,3 @@
-import React from 'react'
-
 const ArticleListing = () => {
     return (
         <div>
